+++ src/pages/Checklist/ChecklistPage.tsx
@@ -5,7 +5,7 @@ import withAccessControl from '../../services/withAccessControl';
 import useCommonHooks from '../../utils/useCommonHooks';
 import { Route, Switch } from 'react-router-dom';
 import ChecklistWrapper from './ChecklistWrapper';
-import NewPunch from './NewPunch/NewPunch';
+import NewPunchWrapper from './NewPunchWrapper';
 import { AsyncStatus } from '../../contexts/McAppContext';
 import { ChecklistResponse, PunchPreview } from '../../services/apiTypes';
 import { Button } from '@equinor/eds-core-react';
@@ -152,7 +152,7 @@ const ChecklistPage = (): JSX.Element => {
                 <Route
                     exact
                     path={`${path}/punch-list/new-punch`}
-                    component={NewPunch}
+                    component={NewPunchWrapper}
                 />
             </Switch>
             <NavigationFooter footerStatus={fetchPunchListStatus}>
